+++ src/sys/ioctl/mod.rs
@@ -232,6 +232,7 @@ pub use self::linux::*;
 
 #[cfg(any(target_os = "dragonfly",
           target_os = "freebsd",
+          target_os = "illumos",
           target_os = "ios",
           target_os = "macos",
           target_os = "netbsd",
@@ -241,6 +242,7 @@ mod bsd;
 
 #[cfg(any(target_os = "dragonfly",
           target_os = "freebsd",
+          target_os = "illumos",
           target_os = "ios",
           target_os = "macos",
           target_os = "netbsd",